+++ lib/routes/youtube/user.ts
@@ -1,12 +1,9 @@
 import { Route, ViewType } from '@/types';
-import cache from '@/utils/cache';
-import utils, { getVideoUrl } from './utils';
 import { config } from '@/config';
-import { parseDate } from '@/utils/parse-date';
-import ofetch from '@/utils/ofetch';
-import * as cheerio from 'cheerio';
 import ConfigNotFoundError from '@/errors/types/config-not-found';
-import NotFoundError from '@/errors/types/not-found';
+import { getDataByUsername as getDataByUsernameYoutubei } from './api/youtubei';
+import { getDataByUsername as getDataByUsernameGoogle } from './api/google';
+import { callApi } from './utils';
 
 export const route: Route = {
     path: '/user/:username/:routeParams?',
@@ -64,84 +61,11 @@ async function handler(ctx) {
     const filterShortsStr = params.get('filterShorts');
     const filterShorts = filterShortsStr === null || filterShortsStr === '' || filterShortsStr === 'true';
 
-    let userHandleData;
-    if (username.startsWith('@')) {
-        userHandleData = await cache.tryGet(`youtube:handle:${username}`, async () => {
-            const link = `https://www.youtube.com/${username}`;
-            const response = await ofetch(link);
-            const $ = cheerio.load(response);
-            const ytInitialData = JSON.parse(
-                $('script')
-                    .text()
-                    .match(/ytInitialData = ({.*?});/)?.[1] || '{}'
-            );
-            const metadataRenderer = ytInitialData.metadata.channelMetadataRenderer;
+    const data = await callApi({
+        googleApi: getDataByUsernameGoogle,
+        youtubeiApi: getDataByUsernameYoutubei,
+        params: { username, embed, filterShorts },
+    });
 
-            const channelId = metadataRenderer.externalId;
-            const channelName = metadataRenderer.title;
-            const image = metadataRenderer.avatar?.thumbnails?.[0]?.url;
-            const description = metadataRenderer.description;
-            const playlistId = (await utils.getChannelWithId(channelId, 'contentDetails', cache)).data.items[0].contentDetails.relatedPlaylists.uploads;
-
-            return {
-                channelName,
-                image,
-                description,
-                playlistId,
-            };
-        });
-    }
-
-    // Get the appropriate playlist ID based on filterShorts setting
-    const playlistId = await (async () => {
-        if (userHandleData?.playlistId) {
-            const origPlaylistId = userHandleData.playlistId;
-
-            return utils.getPlaylistWithShortsFilter(origPlaylistId, filterShorts);
-        } else {
-            const channelData = await utils.getChannelWithUsername(username, 'contentDetails', cache);
-            const items = channelData.data.items;
-
-            if (!items) {
-                throw new NotFoundError(`The channel https://www.youtube.com/user/${username} does not exist.`);
-            }
-
-            const channelId = items[0].id;
-
-            return filterShorts ? utils.getPlaylistWithShortsFilter(channelId, filterShorts) : items[0].contentDetails.relatedPlaylists.uploads;
-        }
-    })();
-
-    const playlistItems = await utils.getPlaylistItems(playlistId, 'snippet', cache);
-    if (!playlistItems) {
-        throw new NotFoundError("This channel doesn't have any content.");
-    }
-
-    return {
-        title: `${userHandleData?.channelName || username} - YouTube`,
-        link: username.startsWith('@') ? `https://www.youtube.com/${username}` : `https://www.youtube.com/user/${username}`,
-        description: userHandleData?.description || `YouTube user ${username}`,
-        image: userHandleData?.image,
-        item: playlistItems.data.items
-            .filter((d) => d.snippet.title !== 'Private video' && d.snippet.title !== 'Deleted video')
-            .map((item) => {
-                const snippet = item.snippet;
-                const videoId = snippet.resourceId.videoId;
-                const img = utils.getThumbnail(snippet.thumbnails);
-                return {
-                    title: snippet.title,
-                    description: utils.renderDescription(embed, videoId, img, utils.formatDescription(snippet.description)),
-                    pubDate: parseDate(snippet.publishedAt),
-                    link: `https://www.youtube.com/watch?v=${videoId}`,
-                    author: snippet.videoOwnerChannelTitle,
-                    image: img.url,
-                    attachments: [
-                        {
-                            url: getVideoUrl(videoId),
-                            mime_type: 'text/html',
-                        },
-                    ],
-                };
-            }),
-    };
+    return data;
 }
